Refactored Column into its own file
